Continued architectural change.
Added Groups, a document does now have a owner, plus maybe some contributors, which can be users or groups.
A user can be part of a group.
ScoreController, to update a score, you must be the owner of one of the contributors, or you must belong to a group that is.