Added Sources Business, now will check that the source exists before creating or updating a score.
For this to work I implemented basic crud for Source like for User.
User username is now the id used in the database.
Update and delete now require id and revision of the document to be changed / deleted.
The management of the id / rev is let to the datamapper, so you can manage id / rev check there. (in this case we simply let couchdb do it)
Fields musicalSourceId, musicalSourcePage, ... now are in two objects, IMusicalSource and ITextualSource in IScore. TextualSource may be a play or either a book atm. (at the moment)